bugfix(img-insert-in-a-anchor): add test and rollback TypeAfterLinkPlugin
